+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `VERSION`: Add inital version
 - `cli/libtman`: Add function to check if task ID exists
+- `src`: Add skeleton for board functions
 
 ### Changed